backend/requirements.txt
fastapi>=0.111.0
uvicorn>=0.30.1
sqlalchemy>=2.0.30
asyncpg>=0.29.0
alembic>=1.13.1
pydantic>=2.7.4
pydantic-settings>=2.3.4
python-multipart>=0.0.9
openpyxl>=3.1.5
PyJWT[crypto]>=2.8.0
httpx>=0.27.0
yfinance>=0.2.54
psycopg2-binary>=2.9.9
schedule>=1.2.2
python-dotenv>=1.0.1
